Move logic to model
