+++ spec/spec_helper.rb
@@ -60,6 +60,7 @@ RSpec.configure do |config|
   end
 
   config.before(:each) do
+    Sidekiq::Worker.clear_all
     DatabaseCleaner.start
   end
